# crates/secure_smoke_service/openapi.yaml
openapi: "3.1.0"
info:
  title: SunLit Security Smoke-Test Service
  description: >
    A purpose-built axum microservice with 54 routes, each exercising a
    specific security control against a known attack class. Includes 15
    mobile security routes covering MASVS controls. Designed for automated
    DAST scanning with OWASP ZAP.
  version: "0.1.0"
  license:
    name: MIT
servers:
  - url: http://127.0.0.1:3001
    description: Local development server
paths:
...
  /smoke/authz/allow:
    get:
      summary: Authorized access granted
      operationId: smokeAuthzAllow
      tags: [authorization]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Access granted with admin role
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Authentication failed
        "403":
          description: Access denied
  /smoke/authz/deny:
    get:
      summary: Unauthorized access denied
      operationId: smokeAuthzDeny
      tags: [authorization]
      security:
        - bearerAuth: []
      responses:
        "403":
          description: Access denied — insufficient role
  /smoke/authz/cross-tenant:
    get:
      summary: Cross-tenant access denied
      operationId: smokeAuthzCrossTenant
      tags: [authorization]
      security:
        - bearerAuth: []
      responses:
        "403":
          description: Access denied — wrong tenant
  /smoke/authz/privilege-escalation:
    post:
      summary: Privilege escalation attempt denied
      operationId: smokeAuthzPrivilegeEscalation
      tags: [authorization]
      security:
        - bearerAuth: []
      responses:
        "403":
          description: Escalation blocked
  /smoke/authz/idor:
    get:
      summary: Insecure direct object reference check
      operationId: smokeAuthzIdor
      tags: [authorization]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Access allowed (ownership matches or policy permits)
        "403":
          description: Access denied

  # ── Data Protection ───────────────────────────────────────────────────
...
  /smoke/events/log-injection:
    post:
      summary: Log injection payload sanitised
      operationId: smokeEventsLogInjection
      tags: [security-events]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogInjectionRequest"
      responses:
        "200":
          description: Event logged safely
          content:
            application/json:
              schema:
                type: object
  /smoke/events/redaction:
    post:
      summary: PII redacted in security events
      operationId: smokeEventsRedaction
      tags: [security-events]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedactionRequest"
      responses:
        "200":
          description: Event emitted with redacted PII
          content:
            application/json:
              schema:
                type: object

  # ── Mobile Security (MASVS) ──────────────────────────────────────────
...